Implement reference counting for devices, contributors, locations & exchanges. (#935)
This pull request introduces a new `reference_count` field across
multiple test cases and updates imports to include `AccountData` where
necessary. These changes ensure consistency in test data structures and
prepare the codebase for future functionality that may depend on the
`reference_count` field.

### Addition of `reference_count` Field:
* Added the `reference_count` field to various test data structures in
files such as `activator_metrics.rs`, `device.rs`, `exchange.rs`,
`location.rs`, `user.rs`, and several files under `smartcontract/cli`.
This change ensures all test cases align with the updated data model.
[[1]](diffhunk://#diff-d2c97b8908f4c2dfc05d08b66e7eeb38241c4e0c628bf2e659dd17ed3364073aR106)
[[2]](diffhunk://#diff-0aed512054e126f7dc8b0bffbc2aad2bd5e2aade35d7e902e08138cc36f94341L109-R115)
[[3]](diffhunk://#diff-b8b6840e7c5c188e05e0dad2007d6f2e7cad9215cff86898ca016d9ee73b430fR31)
[[4]](diffhunk://#diff-61ebc26d8da45511f5934bd93713166a175173d7712ff3f494e17dfc1c7f259dR32)
[[5]](diffhunk://#diff-2107e91f8266f8029aa94b2388d3a6f1d9c0761df5d67c8ea8abd161d1b1183dR476)
[[6]](diffhunk://#diff-7a89574ce930f2bedf4951036cef91d95e34445452d75c567dcd5310f6ed053cL79-R79)
[[7]](diffhunk://#diff-2827ea29f271e58afbebc2d0b44de20e0c4c7cf90a4124415eaa5a5a91c8747cL66-R66)
[[8]](diffhunk://#diff-28204afbc0daa4568757cbf5a6f1073eabc5092155099845b859903fce876ac0R188)
and others)

### Updates to Imports:
* Updated imports in `device.rs` and `link.rs` to include `AccountData`,
reflecting changes in dependencies and ensuring compatibility with the
added `reference_count` field.
[[1]](diffhunk://#diff-0aed512054e126f7dc8b0bffbc2aad2bd5e2aade35d7e902e08138cc36f94341L94-R94)
[[2]](diffhunk://#diff-d9c91ea17e4fe3e26302887a60fd0596f950bcc4949005b71ae4fbf2acfc5ebbL117-R117)

### Mock Client Enhancements:
* Enhanced mock clients in `device.rs` and `link.rs` to include
expectations for retrieving and handling `reference_count` via
`AccountData`. This improves the accuracy of test simulations.
[[1]](diffhunk://#diff-0aed512054e126f7dc8b0bffbc2aad2bd5e2aade35d7e902e08138cc36f94341R159-R167)
[[2]](diffhunk://#diff-d9c91ea17e4fe3e26302887a60fd0596f950bcc4949005b71ae4fbf2acfc5ebbR190-R195)

These changes collectively enhance test coverage and maintain
consistency with the evolving data model.

---------

+++ smartcontract/sdk/rs/src/commands/contributor/delete.rs
@@ -1,4 +1,7 @@
-use crate::{commands::globalstate::get::GetGlobalStateCommand, DoubleZeroClient};
+use crate::{
+    commands::{contributor::get::GetContributorCommand, globalstate::get::GetGlobalStateCommand},
+    DoubleZeroClient,
+};
 use doublezero_serviceability::{
     instructions::DoubleZeroInstruction, processors::contributor::delete::ContributorDeleteArgs,
 };
@@ -14,6 +17,19 @@ impl DeleteContributorCommand {
         let (globalstate_pubkey, _globalstate) = GetGlobalStateCommand {}
             .execute(client)
             .map_err(|_err| eyre::eyre!("Globalstate not initialized"))?;
+
+        let (_, contributor) = GetContributorCommand {
+            pubkey_or_code: self.pubkey.to_string(),
+        }
+        .execute(client)
+        .map_err(|_err| eyre::eyre!("Contributor not found"))?;
+
+        if contributor.reference_count > 0 {
+            return Err(eyre::eyre!(
+                "Contributor cannot be deleted, it has {} references",
+                contributor.reference_count
+            ));
+        }
 
         client.execute_transaction(
             DoubleZeroInstruction::DeleteContributor(ContributorDeleteArgs {}),
@@ -35,9 +51,14 @@ mod tests {
         instructions::DoubleZeroInstruction,
         pda::{get_contributor_pda, get_globalstate_pda},
         processors::contributor::delete::ContributorDeleteArgs,
+        state::{
+            accountdata::AccountData,
+            accounttype::AccountType,
+            contributor::{Contributor, ContributorStatus},
+        },
     };
     use mockall::predicate;
-    use solana_sdk::{instruction::AccountMeta, signature::Signature};
+    use solana_sdk::{instruction::AccountMeta, pubkey::Pubkey, signature::Signature};
 
     #[test]
     fn test_commands_contributor_delete_command() {
@@ -45,6 +66,20 @@ mod tests {
 
         let (globalstate_pubkey, _globalstate) = get_globalstate_pda(&client.get_program_id());
         let (pda_pubkey, _) = get_contributor_pda(&client.get_program_id(), 1);
+        let contributor = Contributor {
+            account_type: AccountType::Contributor,
+            index: 1,
+            bump_seed: 255,
+            code: "cont".to_string(),
+            status: ContributorStatus::Activated,
+            reference_count: 0,
+            owner: Pubkey::default(),
+        };
+
+        client
+            .expect_get()
+            .with(predicate::eq(pda_pubkey))
+            .returning(move |_| Ok(AccountData::Contributor(contributor.clone())));
 
         client
             .expect_execute_transaction()
